Remove alternative image providers temporarily
Those alternatives may work fine, but when the window provider is SDL2, then it simply won't start without `libsdl2-image-2.0-0` present.

As a temporary fix, removing the alternatives makes sure `libsdl2-2.0-0` will be installed.

`python3-pil` certainly works as an image replacement (tested on 16.04, 18.04 py2, py3), the only change that needs to happen is Kivy not requiring that SDL2-image dependency when it's not actually being used. Then the others can be readded.